tataxon_bot/requirements.txt
aioredis==2.0.1
amqp==5.2.0
anyio==4.6.0
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.30.0
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
channels==4.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.16
django-redis==5.4.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
flower==2.0.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
humanize==4.11.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
lxml==5.3.0
pillow==10.4.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psycopg==3.2.3
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.6
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
referencing==0.35.1
rfc3986==1.5.0
rpds-py==0.20.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
